Version 1.14
Vimake,    compile-assemble-link-run-debug a file or files in a vim session with hot keys.
This version of vimake has many improvements over the earlier version. It now can load,build and run
in a vim session lua. It can automatically switch the ddd front end debugger to load the following
file types. Asm, C/CPP, Bash, Python and Perl. It can also run the ruby debugger and the other script languages  in a vim shell. A full installation guide  and tutorial is included in the doc directory. To build this version copy the package to your home directory then run: tar zxvf vimake-1.14.tar.gz.
Change directory to $HOME/vimake-1.14 and follow the README instructions.